test: split reserved-scratch contracts under 300 lines
Extract shared case runner helpers and move ITA/ignoreCase/literal
pathspec edges into separate files so no reserved-scratch subsystem
test exceeds the AGENTS.md single-file growth threshold.

Generated-By: looper 0.11.2 (runner=fixer, agent=grok-build)
